requirements.txt
cryptography
asgiref
autopep8
Django==3.0
pycodestyle
pytz
sqlparse
Unipath
dj-database-url
python-decouple
gunicorn
whitenoise
pyodbc
django-mssql-backend-azure
djangorestframework
markdown
channels
channels_redis
beatserver
django-pandas
balena-sdk

django-celery-beat
#django-timezone-field
#gevent
#celery

django-mqtt-bridge